[ignore] Update pydantic_compat.py to support extra Pydantic methods and classes.
