+++ plugins/module_utils/models/base.py
@@ -9,7 +9,7 @@ from __future__ import absolute_import, division, print_function
 __metaclass__ = type
 
 from abc import ABC
-from ansible_collections.cisco.nd.plugins.module_utils.pydantic_compat import BaseModel, ConfigDict
+from ansible_collections.cisco.nd.plugins.module_utils.common.pydantic_compat import BaseModel, ConfigDict
 from typing import List, Dict, Any, ClassVar, Set, Tuple, Union, Literal, Optional
 from ansible_collections.cisco.nd.plugins.module_utils.utils import issubset
